Fix multiple flaws in this UpdateDB
First: Because it was all in one loop. Maybe a non targetID team was getting attached to the wrong team.

Second: as we weren't doing any moves until after the loop. GetScript was always returning false. So scripts were overriding each other.

Also, as scripts were always attaching to the first team it sees, we need to look for otheres that dont have scripts.